Updated to send tamper firmware
Will now send tamper firmware to stm32 via the USB connection.

stm32 commit
featureTamperFirmware  56b3ac9624c5b911e41543938cba3f9a4a4d683e

libhal commit
featureDiamondKeyTamper c565518e3205dfa00d12fdf383d17bad5f9e1649

+++ hsm_software/sw/safe_shutdown.py
@@ -5,12 +5,16 @@ import os
 
 class SafeShutdown(object):
     """Class to ensure the HSM shuts down correctly"""
-    def __init__(self):
+    def __init__(self, dont_shutdown = False):
         self.callbacks = []
         self.restart_callbacks = []
+        self.dont_shutdown = dont_shutdown
 
     def shutdown(self):        
         self.prepareForShutdown()
+        if (self.dont_shutdown):
+            exit(0)
+
         os.system('sudo shutdown -h now')
 
     def restart(self):  
@@ -20,6 +24,10 @@ class SafeShutdown(object):
 
         # normal shutdown callbacks
         self.prepareForShutdown()
+
+        if (self.dont_shutdown):
+            exit(0)
+
         os.system('sudo reboot')
 
     def prepareForShutdown(self):
